# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for SGMA (Sistema de Gestión de Mantenimiento y Activos)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: Listar activos
      parameters:
        - name: estado
          in: query
          required: false
          schema:
            type: string
        - name: tipo
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Crear activo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Obtener activo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateAsset
      tags: [assets]
      summary: Actualizar activo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    delete:
      operationId: deleteAsset
      tags: [assets]
      summary: Eliminar activo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /assets/{id}/history:
    get:
      operationId: getAssetHistory
      tags: [assets]
      summary: Hoja de vida del equipo (historial de OT)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset work order history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkOrder"
...
  /maintenance-plans:
    get:
      operationId: listMaintenancePlans
      tags: [maintenance-plans]
      summary: Listar planes de mantenimiento preventivo
      responses:
        "200":
          description: List of maintenance plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaintenancePlan"
    post:
      operationId: createMaintenancePlan
      tags: [maintenance-plans]
      summary: Crear plan de mantenimiento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenancePlanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenancePlan"
  /maintenance-plans/{id}:
    patch:
      operationId: updateMaintenancePlan
      tags: [maintenance-plans]
      summary: Actualizar plan de mantenimiento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenancePlanUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenancePlan"
    delete:
      operationId: deleteMaintenancePlan
      tags: [maintenance-plans]
      summary: Eliminar plan de mantenimiento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /locations:
    get:
      operationId: listLocations
      tags: [locations]
      summary: Listar ubicaciones
      responses:
        "200":
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Location"
    post:
      operationId: createLocation
      tags: [locations]
      summary: Crear ubicación
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
  /locations/{id}:
    patch:
      operationId: updateLocation
      tags: [locations]
      summary: Actualizar ubicación
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
    delete:
      operationId: deleteLocation
      tags: [locations]
      summary: Eliminar ubicación
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Asset:
      type: object
      required:
        - id
        - codigo
        - nombre
        - tipo
        - estado
      properties:
        id:
          type: integer
        codigo:
          type: string
        nombre:
          type: string
        tipo:
          type: string
        marca:
          type: ["string", "null"]
        modelo:
          type: ["string", "null"]
        serie:
          type: ["string", "null"]
        anio:
          type: ["integer", "null"]
        ubicacionId:
          type: ["integer", "null"]
        ubicacionNombre:
          type: ["string", "null"]
        centroTrabajoId:
          type: ["integer", "null"]
        centroTrabajoNombre:
          type: ["string", "null"]
        estado:
          type: string
        responsable:
          type: ["string", "null"]
        horometro:
          type: ["number", "null"]
        kilometraje:
          type: ["number", "null"]
        horasAcumuladas:
          type: ["number", "null"]
        vidaUtil:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
        notas:
          type: ["string", "null"]
    AssetInput:
      type: object
      required:
        - codigo
        - nombre
        - tipo
        - estado
      properties:
        codigo:
          type: string
          minLength: 1
        nombre:
          type: string
          minLength: 1
        tipo:
          type: string
        marca:
          type: string
        modelo:
          type: string
        serie:
          type: string
        anio:
          type: integer
        ubicacionId:
          type: integer
        centroTrabajoId:
          type: integer
        estado:
          type: string
        responsable:
          type: string
        horometro:
          type: number
        kilometraje:
          type: number
        horasAcumuladas:
          type: number
        vidaUtil:
          type: integer
        imageUrl:
          type: string
        notas:
          type: string
    AssetUpdate:
      type: object
      properties:
        codigo:
          type: string
          minLength: 1
        nombre:
          type: string
          minLength: 1
        tipo:
          type: string
        marca:
          type: string
        modelo:
          type: string
        serie:
          type: string
        anio:
          type: integer
        ubicacionId:
          type: integer
        centroTrabajoId:
          type: integer
        estado:
          type: string
        responsable:
          type: string
        horometro:
          type: number
        kilometraje:
          type: number
        horasAcumuladas:
          type: number
        vidaUtil:
          type: integer
        imageUrl:
          type: string
        notas:
          type: string
    WorkOrder:
      type: object
      required:
        - id
        - numero
        - equipoId
        - tipo
        - prioridad
        - estado
        - fechaCreacion
      properties:
        id:
          type: integer
        numero:
          type: string
        equipoId:
          type: integer
        equipoNombre:
          type: ["string", "null"]
        equipoCodigo:
          type: ["string", "null"]
        tipo:
          type: string
        prioridad:
          type: string
        estado:
          type: string
        tecnicoId:
          type: ["integer", "null"]
        tecnicoNombre:
          type: ["string", "null"]
        centroTrabajoId:
          type: ["integer", "null"]
        centroTrabajoNombre:
          type: ["string", "null"]
        descripcion:
          type: ["string", "null"]
        reporteFalla:
          type: ["string", "null"]
        diagnostico:
          type: ["string", "null"]
        causaRaiz:
          type: ["string", "null"]
        solucion:
          type: ["string", "null"]
        horasEstimadas:
          type: ["number", "null"]
        horasReales:
          type: ["number", "null"]
        costoManoObra:
          type: ["number", "null"]
        costoRepuestos:
          type: ["number", "null"]
        costoTotal:
          type: ["number", "null"]
        fechaCreacion:
          type: string
        fechaProgramada:
          type: ["string", "null"]
        fechaCierre:
          type: ["string", "null"]
...
    MaintenancePlan:
      type: object
      required:
        - id
        - nombre
        - equipoId
        - tipoFrecuencia
        - intervalo
        - activo
      properties:
        id:
          type: integer
        nombre:
          type: string
        equipoId:
          type: integer
        equipoNombre:
          type: ["string", "null"]
        tipoFrecuencia:
          type: string
        intervalo:
          type: integer
        unidad:
          type: ["string", "null"]
        descripcion:
          type: ["string", "null"]
        proximaFecha:
          type: ["string", "null"]
        proximoHorometro:
          type: ["number", "null"]
        activo:
          type: boolean
    MaintenancePlanInput:
      type: object
      required:
        - nombre
        - equipoId
        - tipoFrecuencia
        - intervalo
      properties:
        nombre:
          type: string
          minLength: 1
        equipoId:
          type: integer
        tipoFrecuencia:
          type: string
        intervalo:
          type: integer
        unidad:
          type: string
        descripcion:
          type: string
        proximaFecha:
          type: string
        proximoHorometro:
          type: number
        activo:
          type: boolean
    MaintenancePlanUpdate:
      type: object
      properties:
        nombre:
          type: string
          minLength: 1
        equipoId:
          type: integer
        tipoFrecuencia:
          type: string
        intervalo:
          type: integer
        unidad:
          type: string
        descripcion:
          type: string
        proximaFecha:
          type: string
        proximoHorometro:
          type: number
        activo:
          type: boolean
...
    Location:
      type: object
      required:
        - id
        - nombre
        - tipo
      properties:
        id:
          type: integer
        nombre:
          type: string
        tipo:
          type: string
        direccion:
          type: ["string", "null"]
        ciudad:
          type: ["string", "null"]
        equiposCount:
          type: ["integer", "null"]
    LocationInput:
      type: object
      required:
        - nombre
        - tipo
      properties:
        nombre:
          type: string
          minLength: 1
        tipo:
          type: string
        direccion:
          type: string
        ciudad:
          type: string
    LocationUpdate:
      type: object
      properties:
        nombre:
          type: string
          minLength: 1
        tipo:
          type: string
        direccion:
          type: string
        ciudad:
          type: string
...