# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for SGMA (Sistema de Gestión de Mantenimiento y Activos)
 * OpenAPI spec version: 0.1.0
 */

export interface Asset {
  id: number;
  codigo: string;
  nombre: string;
  tipo: string;
  /** @nullable */
  marca?: string | null;
  /** @nullable */
  modelo?: string | null;
  /** @nullable */
  serie?: string | null;
  /** @nullable */
  anio?: number | null;
  /** @nullable */
  ubicacionId?: number | null;
  /** @nullable */
  ubicacionNombre?: string | null;
  /** @nullable */
  centroTrabajoId?: number | null;
  /** @nullable */
  centroTrabajoNombre?: string | null;
  estado: string;
  /** @nullable */
  responsable?: string | null;
  /** @nullable */
  horometro?: number | null;
  /** @nullable */
  kilometraje?: number | null;
  /** @nullable */
  horasAcumuladas?: number | null;
  /** @nullable */
  vidaUtil?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  notas?: string | null;
}